BACKEND/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "authlib>=1.6.8",
    "bcrypt==4.0.1",
    "fastapi>=0.133.0",
    "google-auth>=2.48.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langgraph>=1.0.9",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.2",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "pypdf>=6.7.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.0",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.41.0",
]
